requirements.txt
altgraph==0.17.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
et-xmlfile==1.1.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
numpy
openpyxl
pandas
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
Telethon==1.24.0
urllib3==1.26.9
websocket-client==1.2.3
zope.event==4.5.0
zope.interface==5.4.0
